Show dataframe as rich HTML tables
There are a few different issues here:

1. Reticulate does not display pandas data frames correctly.
   It ignores `_repr_html_()` and fails to renders them as HTML at all.
   There is an open issue for this, but no activity.
2. There is not option in Quarto or Knitr
   to display R dataframes as just the first and last few rows
   together with their dimensions.
   This could be fixed via a custom Kable function,
   but it's a bit easier to use rpy2,
   so I'm using that for now, might switch in the future.

The current solution is good enough,
but it is a bit inconvenient to have to indicate the echo and output options
on every cell with dataframe tables.
These bugs should really just be fixed in reticulate/quarto/knitr directly,
but these workarounds are good enough for now.
Maybe I will make the function into a knitr hook
which means there is just a cell option to add instead of the additional
cell.