Switch to using tabbed content instead of margins for packages
I would like to try to write the book in a modular manner if possible.
Ideally,
the theory part would be outside the tabbed content boxes,
and anything related to a specific packages
would go inside of tabbed content
(as much as possible).
This will make it easier to add other (grammar of graphics-based) packages
in a plug and play manner,
without having to rewrite entire theory sections.
I have changed the margin notes to be about optional content.